% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeDataset.R
\name{makeDataset}
\alias{makeDataset}
\title{A function to build cancer dataset using BE1 sparse matrices.}
\usage{
makeDataset(
  input.folder,
  output.folder,
  cell.lines = c("A549", "CCL-185-IG", "CRL5868", "DV90", "HCC78", "HTB178", "PC9",
    "PBMCs"),
  n.cells = c(100, 100, 100, 100, 100, 100, 100, 10)
)
}
\arguments{
\item{input.folder, }{a character string indicating the path where BE1run12.zip from figshare doi:10.6084/m9.figshare.23939481 has been downloaded.}

\item{output.folder, }{a character string indicating the path where to save the data.}

\item{cell.lines, }{a vector indicating the cell lines to be included in the dataset. Seven options:}

\item{n.cells, }{a vector indicating the number of cell for each cell line. If the number of cell exceed the number of available cell in a data set the full dataset is used.}
}
\value{
A 10XGenomics sparse matrix
}
\description{
This function extract from BE1 experiment a set of cells on the basis of the user requirement
}
\examples{
\dontrun{
    #running deDetection
system("wget http://130.192.119.59/public/annotated_setPace_10000_noC5_clustering.output.txt")
system("wget http://130.192.119.59/public/annotated_setPace_10000_noC5.txt.zip")
unzip("annotated_setPace_10000_noC5.txt.zip")
anovaLike(group="docker", file=paste(getwd(),"annotated_setPace_10000_noC5.txt",sep="/"),
       sep="\t", cluster.file="annotated_setPace_10000_noC5_clustering.output.txt", ref.cluster=3,
       logFC.threshold=1, FDR.threshold=0.05, logCPM.threshold=4, plot=TRUE)

}

}
\author{
Raffaele Calogero, raffaele.calogero [at] unito [dot] it, University of Torino, Italy
}
