% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeDataset.R
\name{makeDataset}
\alias{makeDataset}
\title{A function to build cancer dataset using BE1 sparse matrices.}
\usage{
makeDataset(
  input.folder,
  output.folder,
  cell.lines = c("A549", "CCL-185-IG", "CRL5868", "DV90", "HCC78", "HTB178", "PC9",
    "PBMCs"),
  n.cells = c(100, 100, 100, 100, 100, 100, 100, 10)
)
}
\arguments{
\item{input.folder, }{a character string indicating the path where BE1run12.zip from figshare https://figshare.com/articles/dataset/BE1_10XGenomics_count_matrices/23939481 has been downloaded.}

\item{output.folder, }{a character string indicating the path where to save the data.}

\item{cell.lines, }{a vector indicating the cell lines to be included in the dataset. Seven options:}

\item{n.cells, }{a vector indicating the number of cell for each cell line. If the number of cell exceed the number of available cell in a data set the full dataset is used.}
}
\value{
A 10XGenomics sparse matrix
}
\description{
This function extract from BE1 experiment a set of cells on the basis of the user requirement
}
\examples{
\dontrun{

#download https://figshare.com/articles/dataset/BE1_10XGenomics_count_matrices/23939481
#the downloade file is named BE1run12.zip
unzip("BE1run12.zip")
makeDataset(input.folder="/Users/raffaelecalogero/Desktop/BE1run12", output.folder="/Users/raffaelecalogero/Desktop/tmp", cell.lines = c("A549", "CCL-185-IG", "CRL5868", "DV90", "HCC78", "HTB178", "PC9", "PBMCs"), n.cells=c(100,100,100,100,100,100,100,10))

}

}
\author{
Raffaele Calogero, raffaele.calogero [at] unito [dot] it, University of Torino, Italy
}
